+++ tests/linter/test_basic_corner_cases.py
@@ -9,3 +9,8 @@ def test_newlineless_code():
     code = """func foo():
     pass"""
     lint_code(code)
+
+
+def test_docstring():
+    code = '"""\nhello world!\n"""'
+    lint_code(code)