+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Fixed `max-public-methods` linter check disabling (#222)
  - Default regex for names of constants now allows underscore as a prefix to denote private contants (#223)
  - Fixed parsing of files without newline at the end of file ending with comment (#237)
+ - Fixed linter support for docstrings (#233)
 
 ## [4.1.0] 2023-07-06